refactor(tasks): separate committed publication from flow persistence
