+++ src/tasks/task-registry.ts
@@ -39,7 +39,7 @@ export {
   listTasksForRelatedSessionKey,
   resolveTaskForLookupToken,
 } from "./task-registry-query.js";
-export { publishTaskRecordAfterAtomicStore } from "./task-registry-mutation.js";
+export { publishTaskRecordAfterAtomicStore } from "./task-registry-publication.js";
 export { ensureTaskRegistryReady, reloadTaskRegistryFromStore } from "./task-registry-state.js";
 
 if (process.env.VITEST || process.env.NODE_ENV === "test") {
